token

+++ bookservice/src/main/java/com/hoanghuy2209/bookservice/command/event/BookEventsHandler.java
@@ -2,6 +2,8 @@ package com.hoanghuy2209.bookservice.command.event;
 
 import com.hoanghuy2209.bookservice.command.data.Book;
 import com.hoanghuy2209.bookservice.command.data.BookRepository;
+import com.hoanghuy2209.commonservice.event.BookRollBackStatusEvent;
+import com.hoanghuy2209.commonservice.event.BookUpdateStatusEvent;
 import org.axonframework.eventhandling.EventHandler;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,6 +33,22 @@ public class BookEventsHandler {
 
     @EventHandler
     public void on(BookDeleteEvent event){
-        bookRepository.deleteById(event.getBookId());
+        Book book = bookRepository.findById(event.getBookId()).orElseThrow(null);
+        if(book != null){
+            bookRepository.deleteById(event.getBookId());
+        }
+    }
+
+    @EventHandler
+    public void on(BookUpdateStatusEvent event) {
+        Book book = bookRepository.getById(event.getBookId());
+        book.setIsReady(event.getIsReady());
+        bookRepository.save(book);
+    }
+    @EventHandler
+    public void on(BookRollBackStatusEvent event) {
+        Book book = bookRepository.getById(event.getBookId());
+        book.setIsReady(event.getIsReady());
+        bookRepository.save(book);
     }
 }
